examples/redshift-data-warehouse/requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
aws-cdk-lib==2.145.0
cdk-nag
constructs>=10.3.0, <11.0.0
